+++ honey-back/Honeypot/src/main/java/com/honeypot/honeypot/service/WarnMsgService.java
@@ -34,5 +34,5 @@ public interface WarnMsgService {
      * @param tempId
      * @return
      */
-    List<AlarmextendInfo> getDetailInfo(Integer tempId) throws ParseException;
+    List<AlarmextendInfo> getDetailInfo(long tempId) throws ParseException;
 }
